requirements.txt
# Core dependencies
pydantic>=2.0.0
requests>=2.31.0
networkx>=3.0

# Database support
psycopg2-binary>=2.9.9

# LLM providers
litellm>=1.20.0

# UI
streamlit>=1.28.0

# Optional visualization
plotly>=5.17.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0